Atualizado design de página de login e registro
